+++ src/commands/CommandConfig/utils/resolveConfig.ts
@@ -3,18 +3,19 @@ import type {Context} from "#types/types";
 import type {ParsedProfileConfig} from "#config/ParsedProfileConfig";
 import type {ParsedConfig} from "#config/ParsedConfig";
 import {DEFAULT_PROFILE_NAME} from "#lib/defs";
+import {ERROR_MESSAGES} from "#lib/errorMessages";
 
 export function resolveConfig(ctx: Context, options: SharedOptions, mode?: "save"): ParsedProfileConfig
 export function resolveConfig(ctx: Context, options: SharedOptions, mode?: "profile"): ParsedProfileConfig
 export function resolveConfig(ctx: Context, options: SharedOptions, mode: "config-only"): ParsedConfig
 export function resolveConfig(ctx: Context, options: SharedOptions, mode: "profile" | "save" | "config-only" = 'profile'): ParsedProfileConfig | ParsedConfig {
   if (options.local && !ctx.config.local) {
-    ctx.stdOut('No local config found. Run `tunli init` to create one.')
+    ctx.stdOut(ERROR_MESSAGES.NO_LOCAL_CONFIG)
     ctx.exit(1)
   }
 
   if (options.global && !ctx.config.global.exists()) {
-    ctx.stdOut('No global config found. Run `tunli register` to create one.')
+    ctx.stdOut(ERROR_MESSAGES.NO_GLOBAL_CONFIG)
     ctx.exit(1)
   }